add interaction module to SpookyNet layers, test passed
