+++ enerzyme/models/spookynet/interaction.py
@@ -1,9 +1,29 @@
-from typing import Optional
+from typing import Optional, Tuple
 import torch
 from torch.nn import Module, Linear, init
+from ..activation import ACTIVATION_KEY_TYPE
 from ..functional import segment_sum
 from ..layers.electron_embedding import ResidualMLP
+from ..layers.mlp import ResidualStack as _ResidualStack
 from ..layers.attention import Attention
+
+
+def ResidualStack(
+    dim_embedding: int, num_residual: int, 
+    activation_fn: ACTIVATION_KEY_TYPE,
+    use_bias: bool=True, zero_init: bool=True
+) -> _ResidualStack:
+    return _ResidualStack(
+        dim_feature=dim_embedding,
+        num_residual=num_residual,
+        activation_fn=activation_fn,
+        activation_params = {
+            "dim_feature": dim_embedding,
+            "learnable": True
+        },
+        initial_weight1="orthogonal", initial_weight2="zero" if zero_init else "orthogonal",
+        use_bias=use_bias
+    )
 
 
 class LocalInteraction(Module):
@@ -39,21 +59,21 @@ class LocalInteraction(Module):
         num_residual_p: int,
         num_residual_d: int,
         num_residual: int,
-        activation: str = "swish",
+        activation_fn: str = "swish",
     ) -> None:
         """ Initializes the LocalInteraction class. """
         super().__init__()
         self.radial_s = Linear(num_rbf, dim_embedding, bias=False)
         self.radial_p = Linear(num_rbf, dim_embedding, bias=False)
         self.radial_d = Linear(num_rbf, dim_embedding, bias=False)
-        self.resblock_x = ResidualMLP(dim_embedding, num_residual_x, activation)
-        self.resblock_s = ResidualMLP(dim_embedding, num_residual_s, activation)
-        self.resblock_p = ResidualMLP(dim_embedding, num_residual_p, activation)
-        self.resblock_d = ResidualMLP(dim_embedding, num_residual_d, activation)
+        self.resblock_x = ResidualMLP(dim_embedding, num_residual_x, activation_fn)
+        self.resblock_s = ResidualMLP(dim_embedding, num_residual_s, activation_fn)
+        self.resblock_p = ResidualMLP(dim_embedding, num_residual_p, activation_fn)
+        self.resblock_d = ResidualMLP(dim_embedding, num_residual_d, activation_fn)
         self.projection_p = Linear(dim_embedding, 2 * dim_embedding, bias=False)
         self.projection_d = Linear(dim_embedding, 2 * dim_embedding, bias=False)
         self.resblock = ResidualMLP(
-            dim_embedding, num_residual, activation, zero_init=True
+            dim_embedding, num_residual, activation_fn, zero_init=True
         )
         self.reset_parameters()
 
@@ -141,32 +161,26 @@ class NonlocalInteraction(Module):
             'swish': Swish activation function.
             'ssp': Shifted softplus activation function.
     """
-
     def __init__(
         self,
         dim_embedding: int,
         num_residual_q: int,
         num_residual_k: int,
         num_residual_v: int,
-        activation: str = "swish",
+        activation_fn: str = "swish",
     ) -> None:
         """ Initializes the NonlocalInteraction class. """
-        super(NonlocalInteraction, self).__init__()
+        super().__init__()
         self.resblock_q = ResidualMLP(
-            dim_embedding, num_residual_q, activation, zero_init=True
+            dim_embedding, num_residual_q, activation_fn, zero_init=True
         )
         self.resblock_k = ResidualMLP(
-            dim_embedding, num_residual_k, activation, zero_init=True
+            dim_embedding, num_residual_k, activation_fn, zero_init=True
         )
         self.resblock_v = ResidualMLP(
-            dim_embedding, num_residual_v, activation, zero_init=True
+            dim_embedding, num_residual_v, activation_fn, zero_init=True
         )
         self.attention = Attention(dim_embedding, dim_embedding)
-        self.reset_parameters()
-
-    def reset_parameters(self) -> None:
-        """ For compatibility with other modules. """
-        pass
 
     def forward(
         self,
@@ -189,4 +203,122 @@ class NonlocalInteraction(Module):
 
 
 class InteractionModule(Module):
-    ...
+    """
+    InteractionModule of SpookyNet, which computes a single iteration.
+
+    Arguments:
+        num_features (int):
+            Dimensions of feature space.
+        num_basis_functions (int):
+            Number of radial basis functions.
+        num_residual_pre (int):
+            Number of residual blocks applied to atomic features before
+            interaction with neighbouring atoms.
+        num_residual_post (int):
+            Number of residual blocks applied to atomic features after
+            interaction with neighbouring atoms.
+        num_residual_pre_local_i (int):
+            Number of residual blocks applied to atomic features in i branch
+            (central atoms) before computing the local interaction.
+        num_residual_pre_local_j (int):
+            Number of residual blocks applied to atomic features in j branch
+            (neighbouring atoms) before computing the local interaction.
+        num_residual_post_local (int):
+            Number of residual blocks applied to interaction features.
+        num_residual_output (int):
+            Number of residual blocks applied to atomic features in output
+            branch.
+        activation (str):
+            Kind of activation function. Possible values:
+            'swish': Swish activation function.
+            'ssp': Shifted softplus activation function.
+    """
+
+    def __init__(
+        self,
+        dim_embedding: int,
+        num_rbf: int,
+        num_residual_pre: int,
+        num_residual_local_x: int,
+        num_residual_local_s: int,
+        num_residual_local_p: int,
+        num_residual_local_d: int,
+        num_residual_local: int,
+        num_residual_nonlocal_q: int,
+        num_residual_nonlocal_k: int,
+        num_residual_nonlocal_v: int,
+        num_residual_post: int,
+        num_residual_output: int,
+        activation_fn: str = "swish",
+    ) -> None:
+        """ Initializes the InteractionModule class. """
+        super().__init__()
+        # initialize modules
+        self.local_interaction = LocalInteraction(
+            dim_embedding=dim_embedding,
+            num_rbf=num_rbf,
+            num_residual_x=num_residual_local_x,
+            num_residual_s=num_residual_local_s,
+            num_residual_p=num_residual_local_p,
+            num_residual_d=num_residual_local_d,
+            num_residual=num_residual_local,
+            activation_fn=activation_fn,
+        )
+        self.nonlocal_interaction = NonlocalInteraction(
+            dim_embedding=dim_embedding,
+            num_residual_q=num_residual_nonlocal_q,
+            num_residual_k=num_residual_nonlocal_k,
+            num_residual_v=num_residual_nonlocal_v,
+            activation_fn=activation_fn,
+        )
+        self.residual_pre = ResidualStack(dim_embedding, num_residual_pre, activation_fn)
+        self.residual_post = ResidualStack(dim_embedding, num_residual_post, activation_fn)
+        self.resblock = ResidualMLP(
+            dim_embedding, num_residual_output, activation_fn=activation_fn
+        )
+
+    def forward(
+        self,
+        x: torch.Tensor,
+        rbf: torch.Tensor,
+        pij: torch.Tensor,
+        dij: torch.Tensor,
+        idx_i: torch.Tensor,
+        idx_j: torch.Tensor,
+        num_batch: int,
+        batch_seg: torch.Tensor,
+        mask: Optional[torch.Tensor] = None,
+    ) -> Tuple[torch.Tensor, torch.Tensor]:
+        """
+        Evaluate all modules in the block.
+        N: Number of atoms.
+        P: Number of atom pairs.
+        B: Batch size (number of different molecules).
+
+        Arguments:
+            x (FloatTensor [N, num_features]):
+                Latent atomic feature vectors.
+            rbf (FloatTensor [P, num_basis_functions]):
+                Values of the radial basis functions for the pairwise distances.
+            idx_i (LongTensor [P]):
+                Index of atom i for all atomic pairs ij. Each pair must be
+                specified as both ij and ji.
+            idx_j (LongTensor [P]):
+                Same as idx_i, but for atom j.
+            num_batch (int):
+                Batch size (number of different molecules).
+            batch_seg (LongTensor [N]):
+                Index for each atom that specifies to which molecule in the
+                batch it belongs.
+        Returns:
+            x (FloatTensor [N, num_features]):
+                Updated latent atomic feature vectors.
+            y (FloatTensor [N, num_features]):
+                Contribution to output atomic features (environment
+                descriptors).
+        """
+        x = self.residual_pre(x)
+        l = self.local_interaction(x, rbf, pij, dij, idx_i, idx_j)
+        n = self.nonlocal_interaction(x, num_batch, batch_seg, mask)
+        x = self.residual_post(x + l + n)
+        return x, self.resblock(x)
